chore: cleanup in page domain

+++ internal/application/service/span.go
@@ -3,7 +3,7 @@ package service
 type SpanCreateParams struct {
 	PageID    string
 	SessionID string
-	Date      int
+	Date      int64
 	UserAgent string
 	Hostname  string
 	Pathname  string
